fix nav link to be hidden and update access tag (#70607)
