+++ x-pack/plugins/security_solution/common/constants.ts
@@ -35,6 +35,16 @@ export const DEFAULT_TIMEPICKER_QUICK_RANGES = 'timepicker:quickRanges';
 export const NO_ALERT_INDEX = 'no-alert-index-049FC71A-4C2C-446F-9901-37XMC5024C51';
 export const ENDPOINT_METADATA_INDEX = 'metrics-endpoint.metadata-*';
 
+export enum SecurityPageName {
+  alerts = 'alerts',
+  overview = 'overview',
+  hosts = 'hosts',
+  network = 'network',
+  timelines = 'timelines',
+  case = 'case',
+  management = 'management',
+}
+
 export const APP_OVERVIEW_PATH = `${APP_PATH}/overview`;
 export const APP_ALERTS_PATH = `${APP_PATH}/alerts`;
 export const APP_HOSTS_PATH = `${APP_PATH}/hosts`;
