+++ x-pack/plugins/security_solution/public/app/types.ts
@@ -18,16 +18,8 @@ import { State, SubPluginsInitReducer } from '../common/store';
 import { Immutable } from '../../common/endpoint/types';
 import { AppAction } from '../common/store/actions';
 import { TimelineState } from '../timelines/store/timeline/types';
+export { SecurityPageName } from '../../common/constants';
 
-export enum SecurityPageName {
-  alerts = 'alerts',
-  overview = 'overview',
-  hosts = 'hosts',
-  network = 'network',
-  timelines = 'timelines',
-  case = 'case',
-  management = 'management',
-}
 export interface SecuritySubPluginStore<K extends SecuritySubPluginKeyStore, T> {
   initialState: Record<K, T | undefined>;
   reducer: Record<K, Reducer<T, AnyAction>>;
