+++ x-pack/plugins/security_solution/server/lib/timeline/routes/create_timelines_route.ts
@@ -33,7 +33,7 @@ export const createTimelinesRoute = (
         body: buildRouteValidation(createTimelineSchema),
       },
       options: {
-        tags: ['access:siem'],
+        tags: ['access:securitySolution'],
       },
     },
     async (context, request, response) => {
